jarvis-assistant/requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai==1.12.0
pinecone-client==3.0.0
python-dotenv==1.0.0
sentence-transformers==2.2.2
tiktoken==0.5.2
requests==2.31.0
